requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
PyMuPDF==1.23.8
